+++ activerecord/lib/active_record/transactions.rb
@@ -350,9 +350,9 @@ module ActiveRecord
     end
 
     # Add the record to the current transaction so that the #after_rollback and #after_commit callbacks
-    # can be called.
+    # can be called and state restored.
     def add_to_transaction
-      if has_transactional_callbacks?
+      if has_transactional_callbacks? || has_remembered_start_transaction_state?
         self.class.connection.add_transaction_record(self)
       else
         sync_with_transaction_state
@@ -456,6 +456,10 @@ module ActiveRecord
 
     def has_transactional_callbacks? # :nodoc:
       !_rollback_callbacks.empty? || !_commit_callbacks.empty? || !_before_commit_callbacks.empty?
+    end
+
+    def has_remembered_start_transaction_state? # :nodoc:
+      @_start_transaction_state.any?
     end
 
     # Updates the attributes on this particular Active Record object so that