+++ activerecord/test/cases/transactions_test.rb
@@ -176,10 +176,33 @@ class TransactionTest < ActiveRecord::TestCase
   end
 
   def test_transaction_state_is_cleared_when_record_is_persisted
-    author = Author.create! name: 'foo'
+    author = Author.create!(name: 'foo')
     author.name = nil
     assert_not author.save
     assert_not author.new_record?
+  end
+
+  def test_transaction_state_is_restored_when_saving_and_rollback
+    author = Author.new(name: 'foo')
+    Author.transaction do
+      author.save
+      raise ActiveRecord::Rollback
+    end
+    assert_nil author.id
+    assert_equal true, author.instance_variable_get(:'@new_record')
+    assert_equal false, author.instance_variable_get(:'@destroyed')
+  end
+
+  def test_transaction_state_is_restored_when_destroying_and_rollback
+    author = Author.create!(name: 'foo')
+    author_id = author.id
+    Author.transaction do
+      author.destroy
+      raise ActiveRecord::Rollback
+    end
+    assert_equal author_id, author.id
+    assert_equal false, author.instance_variable_get(:'@new_record')
+    assert_equal false, author.instance_variable_get(:'@destroyed')
   end
 
   def test_update_should_rollback_on_failure